+++ data/repository/src/main/java/com/idiot/data/repository/GetEstateDetailRepository.kt
@@ -8,6 +8,6 @@ interface GetEstateDetailRepository {
 
   @WorkerThread
   suspend fun requestGetDetailEstate(
-    estateId: Int
+    estateId: Long
   ) : DetailEstate?
 }
